Add new test: how the event handler is able to stop and start working
Just test it's public interface. The only thing not being tested is the
getFactory() method.

Somehow, the test coverage only seems to be working for UI tests. This
is actually somewhat fortunate, since I never got it to work with UI
tests and only with Unit tests. Later, I will research deeper into how
to get it working for both. Preferably at the same report.